+++ apps/api/codelab/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
 import { UserModule } from '@codelab/ddd/modules/users'
-import { InfrastructureModule } from '@codelab/ddd/shared/infrastructure'
+import { InfrastructureModule } from '@codelab/ddd/shared/framework'
 
 @Module({
   imports: [InfrastructureModule, UserModule],
