+++ libs/ddd/modules/users/core/application/handlers/CreateUserCommandHandler.ts
@@ -2,7 +2,7 @@ import { Inject } from '@nestjs/common'
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs'
 import { classToPlain } from 'class-transformer'
 import { fold } from 'fp-ts/lib/Either'
-import { UserDITokens } from '../../../infrastructure/adapter/UserDITokens'
+import { UserDITokens } from '../../../framework/UserDITokens'
 import { User } from '../../domain/user'
 import { CreateUserCommand } from '../commands/CreateUserCommand'
 import { UserUseCaseDto } from '../useCases/UserUseCaseDto'
