+++ libs/ddd/shared/framework/index.ts
@@ -1,0 +1,1 @@
+export * from './nestjs/infrastructure.module'
