+++ libs/ddd/shared/framework/nestjs/infrastructure.module.ts
@@ -1,0 +1,19 @@
+import { Module, OnApplicationBootstrap, Provider } from '@nestjs/common'
+import { CqrsModule } from '@nestjs/cqrs'
+import { initializeTransactionalContext } from 'typeorm-transactional-cls-hooked'
+import {
+  DatabaseModule,
+  GraphqlModule,
+} from '@codelab/ddd/shared/infrastructure'
+
+const providers: Array<Provider> = []
+
+@Module({
+  imports: [CqrsModule, DatabaseModule, GraphqlModule],
+  providers,
+})
+export class InfrastructureModule implements OnApplicationBootstrap {
+  onApplicationBootstrap() {
+    initializeTransactionalContext()
+  }
+}
