--- libs/ddd/shared/infrastructure/adapter/index.ts
@@ -1,1 +1,0 @@
-export * from './InfrastructureModule'
