--- libs/ddd/shared/infrastructure/config/index.ts
@@ -1,3 +1,0 @@
-export * from './DatabaseConfig'
-export * from './GraphqlConfig'
-export * from './HasuraConfig'
