+++ libs/ddd/shared/infrastructure/graphql/config/GraphqlConfig.ts
