+++ libs/ddd/shared/infrastructure/graphql/config/HasuraConfig.ts
