+++ libs/ddd/shared/infrastructure/graphql/models/IEdge.ts
