+++ libs/ddd/shared/infrastructure/graphql/models/IVertex.ts
