+++ libs/ddd/shared/infrastructure/index.ts
@@ -1,4 +1,2 @@
-export * from './adapter'
-export * from './config'
-export * from './persistence/typeorm/entity'
-export * from './persistence/typeorm/TestTypeOrmModule'
+export * from './persistence/typeorm'
+export * from './graphql/GraphqlModule'
