+++ libs/ddd/shared/infrastructure/persistence/config/DatabaseConfig.ts
