--- libs/ddd/shared/infrastructure/persistence/graphql/index.ts
@@ -1,1 +1,0 @@
-export * from './GraphqlController'
