+++ libs/ddd/shared/infrastructure/persistence/typeorm/index.ts
@@ -1,0 +1,3 @@
+export * from './entity'
+export * from './DatabaseModule'
+export * from './TestDatabaseModule'
